+++ RoiFormulaMaker/Notifications/MakeRingRoiNotification.cs
@@ -8,8 +8,8 @@ namespace RoiFormulaMaker.Notifications
     {
         public string StructureName { get; set; }
         public string BaseStructureName { get; set; }
-        public int InnerMargin { get; set; }
-        public int OuterMargin { get; set; }
+        public double InnerMargin { get; set; }
+        public double OuterMargin { get; set; }
 
         public ObservableCollection<string> StructureNames { get; set; }
         public ObservableCollection<string> ContouredStructureNames { get; set; }
